Fixed several small issues.
