+++ api/types/constants.go
@@ -215,3 +215,53 @@ const (
 	// used only internally
 	VerbRotate = "rotate"
 )
+
+const (
+	// RecordAtNode is the default. Sessions are recorded at Teleport nodes.
+	RecordAtNode = "node"
+
+	// RecordAtProxy enables the recording proxy which intercepts and records
+	// all sessions.
+	RecordAtProxy = "proxy"
+
+	// RecordOff is used to disable session recording completely.
+	RecordOff = "off"
+
+	// RecordAtNodeSync enables the nodes to stream sessions in sync mode
+	// to the auth server
+	RecordAtNodeSync = "node-sync"
+
+	// RecordAtProxySync enables the recording proxy which intercepts and records
+	// all sessions, streams the records synchronously
+	RecordAtProxySync = "proxy-sync"
+)
+
+const (
+	// HostKeyCheckYes is the default. The proxy will check the host key of the
+	// target node it connects to.
+	HostKeyCheckYes = "yes"
+
+	// HostKeyCheckNo is used to disable host key checking. This is a insecure
+	// settings which makes MITM possible with no indications, use with caution.
+	HostKeyCheckNo = "no"
+)
+
+// TunnelType is the type of tunnel.
+type TunnelType string
+
+const (
+	// NodeTunnel is a tunnel where the node connects to the proxy (dial back).
+	NodeTunnel TunnelType = "node"
+
+	// ProxyTunnel is a tunnel where a proxy connects to the proxy (trusted cluster).
+	ProxyTunnel TunnelType = "proxy"
+
+	// AppTunnel is a tunnel where the application proxy dials back to the proxy.
+	AppTunnel TunnelType = "app"
+
+	// KubeTunnel is a tunnel where the kubernetes service dials back to the proxy.
+	KubeTunnel TunnelType = "kube"
+
+	// DatabaseTunnel is a tunnel where a database proxy dials back to the proxy.
+	DatabaseTunnel TunnelType = "db"
+)
